pyproject.toml
[project]
name = "data-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncmy>=0.2.11",
    "cryptography>=46.0.7",
    "elasticsearch[async]>=8,<9",
    "fastapi[standard]>=0.136.0",
    "jieba>=0.42.1",
    "langchain>=1.2.15",
    "langchain-community>=0.1.0",
    "dashscope>=1.20.0",
    "langchain-openai>=1.2.0",
    "langgraph>=1.1.8",
    "loguru>=0.7.3",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.1",
    "sqlalchemy>=2.0.49",
]
